moved parts of response to folder Entity #42
