+++ examples/CreditCard/3d-reserve.php
@@ -80,7 +80,7 @@ elseif ($response instanceof FailureResponse):
     // You should handle them based on the given severity as error, warning or information.
     foreach ($response->getStatusCollection() AS $status) {
         /**
-         * @var $status \Wirecard\PaymentSdk\Status
+         * @var $status \Wirecard\PaymentSdk\Entity\Status
          */
         $severity = ucfirst($status->getSeverity());
         $code = $status->getCode();
