+++ examples/PayPal/notify.php
@@ -36,7 +36,7 @@ if ($notification instanceof SuccessResponse) {
     // In our example we iterate over all errors and echo them out. You should display them as error, warning or information based on the given severity.
     foreach ($notification->getStatusCollection() AS $status) {
         /**
-         * @var $status \Wirecard\PaymentSdk\Status
+         * @var $status \Wirecard\PaymentSdk\Entity\Status
          */
         $severity = ucfirst($status->getSeverity());
         $code = $status->getCode();
